move `SavingPlan` to a separate file
